fix: return type  getPotValue

+++ arduino/WaterChannelSubsystem/src/model/Context.cpp
@@ -53,7 +53,7 @@ bool Context::isUnconnected()
   return unconnected;
 }
 
-int Context::getPotValue()
+float Context::getPotValue()
 {
   return potValue;
 }
